prince_of_persia_bot/requirements.txt
python-telegram-bot
python-telegram-bot[socks]
python-dotenv
requests
